Camera Fixes
Hopefully the last Camera Only update.  I think all the buttons are
working properly.  Added in some code to deal with whether the
ThirdPersonCamera script was on a camera that was attached to the
character or a preset scene camera, so that it doesn't break in the test
zone.
Added walls around the edge of the test zone so things don't fall off
the edge.