More Basic Changes
I made an in game UI, showing pause menu and a controls menu while being
able to load settings and main menu

I also added a health bar that in theory should work but Brian's
complicated un commented stat system was weird.

HOW IT SHOULD WORK:
I set the value of HelathBar slider to MaxValue which is a read only
which works but then when I set the HealthBar.value = CurrentHealth
which is a read only variable that's supposed to get the _health.Current
value from the entity class.

Brian skype me to fix this.