Targeting works.
Brian, I'm too lazy to read through it right now but I will soon. I did
a pretty basic target thing. But It needs to check if the buildings have
health or not. So you need to implement the stat thing you did with the
entity script in the building script. :) It's good for now though I'll
work on it more tomorrow or later tonight if I get bored waiting for
1:00am to register for classes.