Player AI finally perfect :)
Some speed modifications need to be added and a slight delay in roam
update path needs to be added.

Other than that, Ai is pretty much done for now until we start getting
real art assets and working with actual animations. As well as when we
add different enemy troops to take their weight into account and slow
them down or speed them up.

We also need to come up with all the building types and resources for
two civilizations